+++ crates/fuel-core/src/service/adapters.rs
@@ -33,6 +33,17 @@ pub mod relayer;
 pub mod sync;
 pub mod txpool;
 
+#[derive(Debug, Clone)]
+pub struct StaticGasPrice {
+    pub gas_price: u64,
+}
+
+impl StaticGasPrice {
+    pub fn new(gas_price: u64) -> Self {
+        Self { gas_price }
+    }
+}
+
 #[derive(Clone)]
 pub struct PoAAdapter {
     shared_state: Option<fuel_core_poa::service::SharedState>,
@@ -40,24 +51,24 @@ pub struct PoAAdapter {
 
 #[derive(Clone)]
 pub struct TxPoolAdapter {
-    service: TxPoolSharedState<P2PAdapter, Database>,
+    service: TxPoolSharedState<P2PAdapter, Database, StaticGasPrice>,
 }
 
 impl TxPoolAdapter {
-    pub fn new(service: TxPoolSharedState<P2PAdapter, Database>) -> Self {
+    pub fn new(service: TxPoolSharedState<P2PAdapter, Database, StaticGasPrice>) -> Self {
         Self { service }
     }
 }
 
 #[derive(Clone)]
 pub struct TransactionsSource {
-    txpool: TxPoolSharedState<P2PAdapter, Database>,
+    txpool: TxPoolSharedState<P2PAdapter, Database, StaticGasPrice>,
     _block_height: BlockHeight,
 }
 
 impl TransactionsSource {
     pub fn new(
-        txpool: TxPoolSharedState<P2PAdapter, Database>,
+        txpool: TxPoolSharedState<P2PAdapter, Database, StaticGasPrice>,
         block_height: BlockHeight,
     ) -> Self {
         Self {