+++ crates/services/producer/src/config.rs
@@ -4,6 +4,5 @@ use fuel_core_types::fuel_types::ContractId;
 pub struct Config {
     pub utxo_validation: bool,
     pub coinbase_recipient: Option<ContractId>,
-    pub gas_price: u64,
     pub metrics: bool,
 }