+++ crates/services/producer/src/mocks.rs
@@ -110,13 +110,14 @@ impl AsRef<MockDb> for MockDb {
     }
 }
 
-fn to_block(component: Components<Vec<ArcPoolTx>>) -> Block {
+fn to_block(component: &Components<Vec<ArcPoolTx>>) -> Block {
     let transactions = component
         .transactions_source
+        .clone()
         .into_iter()
         .map(|tx| tx.as_ref().into())
         .collect();
-    Block::new(component.header_to_produce, transactions, &[])
+    Block::new(component.header_to_produce.clone(), transactions, &[])
 }
 
 impl Executor<Vec<ArcPoolTx>> for MockExecutor {
@@ -124,7 +125,7 @@ impl Executor<Vec<ArcPoolTx>> for MockExecutor {
         &self,
         component: Components<Vec<ArcPoolTx>>,
     ) -> ExecutorResult<UncommittedResult<Changes>> {
-        let block = to_block(component);
+        let block = to_block(&component);
         // simulate executor inserting a block
         let mut block_db = self.0.blocks.lock().unwrap();
         block_db.insert(
@@ -155,7 +156,7 @@ impl Executor<Vec<ArcPoolTx>> for FailingMockExecutor {
         if let Some(err) = err.take() {
             Err(err)
         } else {
-            let block = to_block(component);
+            let block = to_block(&component);
             Ok(UncommittedResult::new(
                 ExecutionResult {
                     block,
@@ -165,6 +166,38 @@ impl Executor<Vec<ArcPoolTx>> for FailingMockExecutor {
                 },
                 Default::default(),
             ))
+        }
+    }
+}
+
+#[derive(Clone)]
+pub struct MockExecutorWithCapture {
+    pub captured: Arc<Mutex<Option<Components<Vec<ArcPoolTx>>>>>,
+}
+
+impl Executor<Vec<ArcPoolTx>> for MockExecutorWithCapture {
+    fn execute_without_commit(
+        &self,
+        component: Components<Vec<ArcPoolTx>>,
+    ) -> ExecutorResult<UncommittedResult<Changes>> {
+        let block = to_block(&component);
+        *self.captured.lock().unwrap() = Some(component);
+        Ok(UncommittedResult::new(
+            ExecutionResult {
+                block,
+                skipped_transactions: vec![],
+                tx_status: vec![],
+                events: vec![],
+            },
+            Default::default(),
+        ))
+    }
+}
+
+impl Default for MockExecutorWithCapture {
+    fn default() -> Self {
+        Self {
+            captured: Arc::new(Mutex::new(None)),
         }
     }
 }