+++ crates/services/txpool/src/config.rs
@@ -62,8 +62,6 @@ pub struct Config {
     pub max_tx: usize,
     /// max depth of connected UTXO excluding contracts
     pub max_depth: usize,
-    /// The minimum allowed gas price
-    pub min_gas_price: u64,
     /// Flag to disable utxo existence and signature checks
     pub utxo_validation: bool,
     /// chain config
@@ -82,7 +80,6 @@ impl Default for Config {
     fn default() -> Self {
         let max_tx = 4064;
         let max_depth = 10;
-        let min_gas_price = 0;
         let utxo_validation = true;
         let metrics = false;
         // 5 minute TTL
@@ -92,7 +89,6 @@ impl Default for Config {
             max_tx,
             max_depth,
             ChainConfig::default(),
-            min_gas_price,
             utxo_validation,
             metrics,
             transaction_ttl,
@@ -108,7 +104,6 @@ impl Config {
         max_tx: usize,
         max_depth: usize,
         chain_config: ChainConfig,
-        min_gas_price: u64,
         utxo_validation: bool,
         metrics: bool,
         transaction_ttl: Duration,
@@ -120,7 +115,6 @@ impl Config {
         Self {
             max_tx,
             max_depth,
-            min_gas_price,
             utxo_validation,
             chain_config,
             metrics,