+++ crates/services/txpool/src/service/test_helpers.rs
@@ -8,6 +8,7 @@ use fuel_core_services::{
     stream::BoxStream,
     Service as ServiceTrait,
 };
+use fuel_core_txpool::types::GasPrice;
 use fuel_core_types::{
     blockchain::SealedBlock,
     entities::coins::coin::Coin,
@@ -32,9 +33,32 @@ use std::cell::RefCell;
 type GossipedTransaction = GossipData<Transaction>;
 
 pub struct TestContext {
-    pub(crate) service: Service<MockP2P, MockDBProvider>,
+    pub(crate) service: Service<MockP2P, MockDBProvider, MockTxPoolGasPrice>,
     mock_db: MockDb,
     rng: RefCell<StdRng>,
+}
+
+#[derive(Debug, Clone)]
+pub struct MockTxPoolGasPrice {
+    pub gas_price: Option<GasPrice>,
+}
+
+impl MockTxPoolGasPrice {
+    pub fn new(gas_price: GasPrice) -> Self {
+        Self {
+            gas_price: Some(gas_price),
+        }
+    }
+
+    pub fn new_none() -> Self {
+        Self { gas_price: None }
+    }
+}
+
+impl GasPriceProviderConstraint for MockTxPoolGasPrice {
+    fn gas_price(&self, _block_height: BlockHeight) -> Option<GasPrice> {
+        self.gas_price
+    }
 }
 
 impl TestContext {
@@ -42,7 +66,7 @@ impl TestContext {
         TestContextBuilder::new().build_and_start().await
     }
 
-    pub fn service(&self) -> &Service<MockP2P, MockDBProvider> {
+    pub fn service(&self) -> &Service<MockP2P, MockDBProvider, MockTxPoolGasPrice> {
         &self.service
     }
 
@@ -188,6 +212,7 @@ impl TestContextBuilder {
 
     pub fn build(self) -> TestContext {
         let rng = RefCell::new(self.rng);
+        let gas_price = 0;
         let config = self.config.unwrap_or_default();
         let mock_db = self.mock_db;
 
@@ -202,6 +227,7 @@ impl TestContextBuilder {
         let importer = self
             .importer
             .unwrap_or_else(|| MockImporter::with_blocks(vec![]));
+        let gas_price_provider = MockTxPoolGasPrice::new(gas_price);
 
         let service = new_service(
             config,
@@ -209,6 +235,7 @@ impl TestContextBuilder {
             importer,
             p2p,
             Default::default(),
+            gas_price_provider,
         );
 
         TestContext {